Refactor memory command to use stub implementation in minimal build
