+++ crates/cli/src/commands/mod.rs
@@ -2,7 +2,7 @@ pub mod gpu;
 #[cfg(not(feature = "minimal"))]
 pub mod memory;
 #[cfg(feature = "minimal")]
-pub mod memory_min;
+pub mod memory_stub;
 pub mod models;
 pub mod tools;
 pub mod smart;
@@ -12,7 +12,7 @@ pub use gpu::GpuCommand;
 #[cfg(not(feature = "minimal"))]
 pub use memory::MemoryCommand;
 #[cfg(feature = "minimal")]
-pub use memory_min::MemoryCommand;
+pub use memory_stub::MemoryCommand;
 pub use models::ModelsCommand;
 pub use tools::ToolsCommand;
 pub use smart::SmartCommand;
